2.9.6 remove requirement for chromosomes in processTags
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/segmentSeq@126630 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/plotGenome.R
@@ -71,7 +71,7 @@ function(aD, loci, chr = 1, limits = c(0, 1e4), samples = NULL, plotType = "pile
                            rectcpu <- matrix(cbind(start(cpu), uu - 0.45, end(cpu), runValue(cpu)), ncol = 4)
                            rectcpu[,4] <- sapply(rectcpu[,4], min, cap)
                            rectcpu <- rectcpu[rectcpu[,4] > 0,,drop = FALSE]                           
-                           rectcpu[,4] <- rectcpu[,4] / sum(values(sTags)$count[tags]) * as.integer(pAD@data[ii,samples[uu]])
+                           rectcpu[,4] <- rectcpu[,4] / sum(values(sTags)$count[tags]) * as.numeric(pAD@data[ii,samples[uu]])
                            if(any(rectcpu[,4] <= 0)) {
                              warning("There are non-zero count tags whose summed count is zero - something's gone wrong with the input data.")
                              rectcpu <- rectcpu[rectcpu[,4] > 0,,drop = FALSE]
